+++ lib/gif_view.dart
@@ -1,51 +1,72 @@
+library gif_view;
+
 import 'dart:async';
-import 'dart:math';
-// ignore: unnecessary_import
-import 'dart:typed_data';
-import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:gif_view/src/gif_controller.dart';
-import 'package:gif_view/src/git_frame.dart';
-import 'package:http/http.dart' as http;
+import 'package:gif_view/src/gif_frame_builder.dart';
 
+import 'src/gif_cache_provider.dart';
+import 'src/gif_loader.dart';
+
+export 'package:gif_view/src/gif_cache_provider.dart';
 export 'package:gif_view/src/gif_controller.dart';
 
-///
-/// Created by
-///
-/// ─▄▀─▄▀
-/// ──▀──▀
-/// █▀▀▀▀▀█▄
-/// █░░░░░█─█
-/// ▀▄▄▄▄▄▀▀
-///
-/// Rafaelbarbosatec
-/// on 23/09/21
-
-final Map<String, List<GifFrame>> _cache = {};
-
 class GifView extends StatefulWidget {
+  /// Optional controller to manage GIF playback externally.
   final GifController? controller;
+
+  /// Duration between frames in the GIF animation.
+  /// Defaults is the fps original file
   final int? frameRate;
   final ImageProvider image;
+
+  /// The height of the GIF view widget.
+  /// If null, the widget will use its parent's height constraints.
   final double? height;
+
+  /// The width of the GIF view widget.
+  /// If null, the widget will use its parent's width constraints.
   final double? width;
-  final Widget? progress;
+  final Widget Function(
+          BuildContext context, Exception error, VoidCallback tryAgain)?
+      errorBuilder;
+  final WidgetBuilder? progressBuilder;
+
+  /// How to fit the image within its bounds.
+  /// Uses Flutter's BoxFit enum.
+  /// Defaults to BoxFit.contain.
   final BoxFit? fit;
   final Color? color;
   final BlendMode? colorBlendMode;
   final AlignmentGeometry alignment;
-  final ImageRepeat repeat;
+  final ImageRepeat imageRepeat;
   final Rect? centerSlice;
   final bool matchTextDirection;
   final bool invertColors;
-  final bool withOpacityAnimation;
   final FilterQuality filterQuality;
   final bool isAntiAlias;
-  final Widget Function(Exception error)? onError;
+
+  final bool withOpacityAnimation;
   final Duration? fadeDuration;
+
+  /// Controls whether the GIF animation should auto-start.
+  /// Defaults to true.
+  final bool autoPlay;
+
+  /// Determines if the GIF should loop continuously.
+  /// When false, the animation will play once and stop.
+  /// Defaults to true.
+  final bool loop;
+  final bool playInverted;
+  final void Function()? onStart;
+
+  /// Callback function that triggers when animation completes.
+  /// Only called when [loop] is false.
+  final void Function()? onFinish;
+  final void Function(int frame)? onFrame;
+  final void Function(int totalFrames)? onLoaded;
 
   GifView.network(
     String url, {
@@ -54,22 +75,29 @@ class GifView extends StatefulWidget {
     this.frameRate,
     this.height,
     this.width,
-    this.progress,
+    this.progressBuilder,
     this.fit,
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
     this.filterQuality = FilterQuality.low,
     this.isAntiAlias = false,
     this.withOpacityAnimation = true,
-    this.onError,
+    this.errorBuilder,
     this.fadeDuration,
     double scale = 1.0,
     Map<String, String>? headers,
+    this.onStart,
+    this.onFinish,
+    this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = NetworkImage(url, scale: scale, headers: headers),
         super(key: key);
 
@@ -80,22 +108,29 @@ class GifView extends StatefulWidget {
     this.frameRate,
     this.height,
     this.width,
-    this.progress,
+    this.progressBuilder,
     this.fit,
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
     this.filterQuality = FilterQuality.low,
     this.isAntiAlias = false,
     this.withOpacityAnimation = true,
-    this.onError,
+    this.errorBuilder,
     this.fadeDuration,
     String? package,
     AssetBundle? bundle,
+    this.onStart,
+    this.onFinish,
+    this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = AssetImage(asset, package: package, bundle: bundle),
         super(key: key);
 
@@ -103,24 +138,31 @@ class GifView extends StatefulWidget {
     Uint8List bytes, {
     Key? key,
     this.controller,
-    this.frameRate = 15,
+    this.frameRate,
     this.height,
     this.width,
-    this.progress,
+    this.progressBuilder,
     this.fit,
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
     this.filterQuality = FilterQuality.low,
     this.isAntiAlias = false,
     this.withOpacityAnimation = true,
-    this.onError,
+    this.errorBuilder,
     this.fadeDuration,
     double scale = 1.0,
+    this.onStart,
+    this.onFinish,
+    this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = MemoryImage(bytes, scale: scale),
         super(key: key);
 
@@ -128,30 +170,66 @@ class GifView extends StatefulWidget {
     Key? key,
     required this.image,
     this.controller,
-    this.frameRate = 15,
+    this.frameRate,
     this.height,
     this.width,
-    this.progress,
+    this.progressBuilder,
     this.fit,
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
     this.filterQuality = FilterQuality.low,
     this.isAntiAlias = false,
     this.withOpacityAnimation = true,
-    this.onError,
+    this.errorBuilder,
     this.fadeDuration,
+    this.onStart,
+    this.onFinish,
+    this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   }) : super(key: key);
+
+  /// Pre-fetches GIF images and stores them in cache for faster loading.
+  ///
+  /// [image] A ImageProvider to be pre-fetched.
+  ///
+  /// Example:
+  /// ```dart
+  /// await GifView.preFetch(AssetImage());
+  /// ```
+  static Future<void> preFetchImage(ImageProvider image) async {
+    await GifLoader.instance.fetch(image);
+  }
+
+  static Future<void> clearCache() async {
+    await GifLoader.instance.clearCache();
+  }
+
+  /// Sets a custom cache provider for GIF image storage.
+  ///
+  /// [provider] An implementation of [GifCacheProvider] to handle caching.
+  /// If null, reverts to the default cache provider.
+  ///
+  /// Example:
+  /// ```dart
+  /// GifView.setCacheProvider(MyCustomCacheProvider());
+  /// ```
+  static void setCacheProvider(GifCacheProvider? provider) {
+    GifLoader.instance.setCacheProvider(provider);
+  }
 
   @override
   GifViewState createState() => GifViewState();
 }
 
-class GifViewState extends State<GifView> with TickerProviderStateMixin {
+class GifViewState extends State<GifView> with SingleTickerProviderStateMixin {
   late GifController controller;
 
   AnimationController? _animationController;
@@ -165,6 +243,14 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
       );
     }
     controller = widget.controller ?? GifController();
+    controller.init(
+      autoPlay: widget.autoPlay,
+      inverted: widget.playInverted,
+      loop: widget.loop,
+      onStart: widget.onStart,
+      onFinish: widget.onFinish,
+      onFrame: widget.onFrame,
+    );
     controller.addListener(_listener);
     WidgetsBinding.instance.addPostFrameCallback(
       (_) => _loadImage(),
@@ -174,8 +260,13 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
 
   @override
   void dispose() {
-    controller.stop();
     controller.removeListener(_listener);
+    if (widget.controller == null) {
+      controller.dispose();
+    } else {
+      controller.stop();
+    }
+
     _animationController?.dispose();
     _animationController = null;
     super.dispose();
@@ -196,12 +287,16 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
       return SizedBox(
         width: widget.width,
         height: widget.height,
-        child: widget.progress,
+        child: widget.progressBuilder?.call(context),
       );
     }
 
     if (controller.status == GifStatus.error) {
-      final errorWidget = widget.onError?.call(controller.exception!);
+      final errorWidget = widget.errorBuilder?.call(
+        context,
+        controller.exception!,
+        () => _loadImage(),
+      );
       if (errorWidget == null) {
         throw controller.exception!;
       }
@@ -221,7 +316,7 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
       color: widget.color,
       colorBlendMode: widget.colorBlendMode,
       alignment: widget.alignment,
-      repeat: widget.repeat,
+      repeat: widget.imageRepeat,
       centerSlice: widget.centerSlice,
       matchTextDirection: widget.matchTextDirection,
       invertColors: widget.invertColors,
@@ -231,49 +326,27 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
     );
   }
 
-  String _getKeyImage(ImageProvider provider) {
-    return provider is NetworkImage
-        ? provider.url
-        : provider is AssetImage
-            ? provider.assetName
-            : provider is MemoryImage
-                ? provider.bytes.toString().substring(0, 100)
-                : provider is FileImage
-                    ? provider.file.path
-                    : Random().nextDouble().toString();
-  }
-
-  Future<List<GifFrame>> _fetchGif(ImageProvider provider) async {
-    List<GifFrame> frameList = [];
-    try {
-      String key = _getKeyImage(provider);
-
-      if (_cache.containsKey(key)) {
-        frameList = _cache[key]!;
-        return frameList;
-      }
-
-      Uint8List? data = await _loadImageBytes(provider);
-
-      if (data == null) {
-        return [];
-      }
-
-      frameList.addAll(await _buildFrames(data));
-
-      _cache.putIfAbsent(key, () => frameList);
-    } catch (e) {
-      controller.error(e as Exception);
-    }
-    return frameList;
-  }
-
   FutureOr _loadImage({bool updateFrames = false}) async {
     controller.loading();
-    final frames = await _fetchGif(widget.image);
-    if (frames.isNotEmpty) {
-      controller.configure(frames, updateFrames: updateFrames);
-      _animationController?.forward(from: 0);
+    try {
+      final data = await GifLoader.instance.fetch(widget.image);
+      if (data != null) {
+        final frames = await GifFrameBuilder(
+          data: data,
+          frameRate: widget.frameRate,
+        ).build();
+        if (frames.isNotEmpty) {
+          widget.onLoaded?.call(frames.length);
+          controller.configure(frames, updateFrames: updateFrames);
+          _animationController?.forward(from: 0);
+        } else {
+          controller.error(Exception('Can not load image'));
+        }
+      } else {
+        controller.error(Exception('Can not load image'));
+      }
+    } catch (e) {
+      controller.error(e as Exception);
     }
   }
 
@@ -281,50 +354,5 @@ class GifViewState extends State<GifView> with TickerProviderStateMixin {
     if (mounted) {
       setState(() {});
     }
-  }
-
-  Future<Uint8List?> _loadImageBytes(ImageProvider<Object> provider) {
-    if (provider is NetworkImage) {
-      final Uri resolved = Uri.base.resolve(provider.url);
-      return http
-          .get(resolved, headers: provider.headers)
-          .then((value) => value.bodyBytes);
-    } else if (provider is AssetImage) {
-      return provider.obtainKey(const ImageConfiguration()).then(
-        (value) async {
-          final d = await value.bundle.load(value.name);
-          return d.buffer.asUint8List();
-        },
-      );
-    } else if (provider is FileImage) {
-      return provider.file.readAsBytes();
-    } else if (provider is MemoryImage) {
-      return Future.value(provider.bytes);
-    }
-    return Future.value(null);
-  }
-
-  Future<Iterable<GifFrame>> _buildFrames(Uint8List data) async {
-    Codec codec = await instantiateImageCodec(
-      data,
-      allowUpscaling: false,
-    );
-
-    List<GifFrame> list = [];
-
-    for (int i = 0; i < codec.frameCount; i++) {
-      FrameInfo frameInfo = await codec.getNextFrame();
-      Duration duration = frameInfo.duration;
-      if (widget.frameRate != null) {
-        duration = Duration(milliseconds: (1000 / widget.frameRate!).ceil());
-      }
-      list.add(
-        GifFrame(
-          ImageInfo(image: frameInfo.image),
-          duration,
-        ),
-      );
-    }
-    return list;
   }
 }